# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Flujo de Caja",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/divisas/tipos-cambio": {
      "get": {
        "tags": [
          "divisas"
        ],
        "summary": "List Tipos Cambio",
        "operationId": "list_tipos_cambio_api_v1_divisas_tipos_cambio_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "moneda_origen",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Moneda Origen"
            }
          },
          {
            "name": "moneda_destino",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Moneda Destino"
            }
          },
          {
            "name": "incluir_inactivos",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Incluir Inactivos"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TipoCambioRead"
                  },
                  "title": "Response List Tipos Cambio Api V1 Divisas Tipos Cambio Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "divisas"
        ],
        "summary": "Create Tipo Cambio",
        "operationId": "create_tipo_cambio_api_v1_divisas_tipos_cambio_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TipoCambioCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TipoCambioRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/divisas/tipos-cambio/{tipo_cambio_id}": {
      "patch": {
        "tags": [
          "divisas"
        ],
        "summary": "Update Tipo Cambio",
        "operationId": "update_tipo_cambio_api_v1_divisas_tipos_cambio__tipo_cambio_id__patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tipo_cambio_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tipo Cambio Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TipoCambioUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TipoCambioRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "divisas"
        ],
        "summary": "Delete Tipo Cambio",
        "operationId": "delete_tipo_cambio_api_v1_divisas_tipos_cambio__tipo_cambio_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tipo_cambio_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tipo Cambio Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TipoCambioRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TipoCambioCreate": {
        "properties": {
          "moneda_origen": {
            "type": "string",
            "maxLength": 3,
            "minLength": 3,
            "title": "Moneda Origen"
          },
          "moneda_destino": {
            "type": "string",
            "maxLength": 3,
            "minLength": 3,
            "title": "Moneda Destino"
          },
          "tasa": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Tasa"
          },
          "fecha_vigencia": {
            "type": "string",
            "format": "date",
            "title": "Fecha Vigencia"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "moneda_origen",
          "moneda_destino",
          "tasa",
          "fecha_vigencia"
        ],
        "title": "TipoCambioCreate"
      },
      "TipoCambioRead": {
        "properties": {
          "moneda_origen": {
            "type": "string",
            "maxLength": 3,
            "minLength": 3,
            "title": "Moneda Origen"
          },
          "moneda_destino": {
            "type": "string",
            "maxLength": 3,
            "minLength": 3,
            "title": "Moneda Destino"
          },
          "tasa": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Tasa"
          },
          "fecha_vigencia": {
            "type": "string",
            "format": "date",
            "title": "Fecha Vigencia"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "deleted_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deleted At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "created_by_id": {
            "type": "string",
            "format": "uuid",
            "title": "Created By Id"
          },
          "updated_by_id": {
            "type": "string",
            "format": "uuid",
            "title": "Updated By Id"
          }
        },
        "type": "object",
        "required": [
          "moneda_origen",
          "moneda_destino",
          "tasa",
          "fecha_vigencia",
          "id",
          "is_active",
          "deleted_at",
          "created_at",
          "updated_at",
          "created_by_id",
          "updated_by_id"
        ],
        "title": "TipoCambioRead"
      },
      "TipoCambioUpdate": {
        "properties": {
          "moneda_origen": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 3,
                "minLength": 3
              },
              {
                "type": "null"
              }
            ],
            "title": "Moneda Origen"
          },
          "moneda_destino": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 3,
                "minLength": 3
              },
              {
                "type": "null"
              }
            ],
            "title": "Moneda Destino"
          },
          "tasa": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tasa"
          },
          "fecha_vigencia": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fecha Vigencia"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "TipoCambioUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/login/access-token"
          }
        }
      }
    }
  }
}